requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
atlassian-python-api==3.41.16
attrs==24.2.0
auth0-python==4.7.2
autogen-agentchat==0.4.3
autogen-core==0.4.3
autogen-ext==0.4.3
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
boto3==1.35.14
botocore==1.35.14
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
cleo==2.1.0
click==8.1.7
clickhouse-connect==0.8.5
cloudpathlib==0.19.0
cohere==5.9.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
crashtest==0.4.1
crewai==0.86.0
crewai-tools==0.17.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
deepeval==1.1.6
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
docx2txt==0.8
dotenv-python==0.0.1
dulwich==0.21.7
durationpy==0.9
embedchain==0.1.126
enum34==1.1.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
faiss-cpu==1.8.0.post1
fastapi==0.115.6
fastavro==1.9.7
fastjsonschema==2.21.1
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.1.1
groq==0.11.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
hyperframe==6.0.1
idna==3.8
importlib-metadata==7.0.0
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
instructor==1.7.2
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.34.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keyring==24.3.1
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.13
langchain-cohere==0.3.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-ollama==0.1.3
langchain-openai==0.2.14
langchain-pinecone==0.0.1
langchain-text-splitters==0.3.4
langcodes==3.4.0
langgraph==0.2.61
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.1.147
langtrace-python-sdk==3.3.14
language_data==1.2.0
litellm==1.56.4
llama-cloud==0.0.17
llama-index==0.11.7
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.0
llama-index-core==0.11.7
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.2
lz4==4.3.3
Mako==1.3.8
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.38
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.20.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.4
onnxruntime==1.20.1
openai==1.58.1
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pathlib==1.0.1
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==11.1.0
pinecone==5.4.2
pinecone-client==3.2.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pip==24.3.1
pip-upgrader==1.4.15
pkginfo==1.12.0
platformdirs==4.3.1
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
portalocker==2.10.1
posthog==3.7.4
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydantic-settings==2.7.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pylance==0.20.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.1
pyparsing==3.1.4
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.391
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.2
ragas==0.1.16
rank-bm25==0.2.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.3.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.27.1
sentence-transformers==3.0.1
sentry-sdk==2.19.2
serpapi==0.1.5
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
simsimd==3.9.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.2
tabulate==0.9.0
tenacity==9.0.0
terminaltables==3.1.10
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
tomlkit==0.13.2
torch==2.2.2
tornado==6.4.1
tqdm==4.66.5
trace-attributes==7.1.0
traitlets==5.14.3
transformers==4.44.2
trio==0.28.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
typer==0.12.5
types-requests==2.32.0.20240907
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uv==0.5.13
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
wasabi==1.1.3
watchfiles==1.0.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xattr==1.1.0
xxhash==3.5.0
yarl==1.10.0
youtube-transcript-api==0.6.3
zipp==3.20.1
zstandard==0.23.0